refactor(images): created new folders to house similar images and correct the new reference path in components and pages
