+++ src/pages/About.tsx
@@ -2,7 +2,7 @@ import ImageGallery from "react-image-gallery";
 import "react-image-gallery/styles/css/image-gallery.css";
 import "../styles/about-gallery-custom.css";
 
-import ingvarSitting from "../images/IngvarSitting.jpg";
+import ingvarSitting from "../images/AasNordreTolerud/IngvarSitting.jpg";
 import "../styles/about.css";
 import { useTranslation } from "react-i18next";
 import ScrollBtn from "../components/ScrollBtn";
