+++ src/pages/Contact.tsx
@@ -1,4 +1,4 @@
-import gingerHouseBack from "../images/GingerHouseBack.jpg";
+import gingerHouseBack2 from "../images/AasNordreTolerud/GingerHouseBack2.png";
 import "../styles/contact.css";
 import ContactForm from "../components/ContactForm";
 import ScrollBtn from "../components/ScrollBtn";
@@ -12,7 +12,7 @@ export default function Contact() {
       <div className="contact-container">
         <img
           className="contact-image"
-          src={gingerHouseBack}
+          src={gingerHouseBack2}
           alt="Ginger House Back View"
         />
       </div>
